Add Turkish translation
I know that this is very scuffed, but this was the most efficient way of implementing that I could think of. As a matter of fact, it was necessary so that Turkish Grammar rules are not broken. (As a native Turkish speaker, I can't afford to break them).

please forgive me